src/main/webapp/app/entities/entity-with-dto/update/entity-with-dto-update.component.spec.ts
jest.mock('@angular/router');

import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { of } from 'rxjs';

import { EntityWithDTOService } from '../service/entity-with-dto.service';
import { EntityWithDTO } from '../entity-with-dto.model';

import { EntityWithDTOUpdateComponent } from './entity-with-dto-update.component';

describe('Component Tests', () => {
  describe('EntityWithDTO Management Update Component', () => {
    let comp: EntityWithDTOUpdateComponent;
    let fixture: ComponentFixture<EntityWithDTOUpdateComponent>;
    let service: EntityWithDTOService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [EntityWithDTOUpdateComponent],
        providers: [FormBuilder, ActivatedRoute],
      })
        .overrideTemplate(EntityWithDTOUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(EntityWithDTOUpdateComponent);
      comp = fixture.componentInstance;
      service = TestBed.inject(EntityWithDTOService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new EntityWithDTO('123');
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new EntityWithDTO();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
